c-testsuite --int=32: float→int conv, libc widths, %lX padding
- codegen: float→int conversion when target is "long" (which captures
  plain int under --int=32) — was missing for both function args and
  variable initializers, so 'intfunc(99.0)' and 'int e = 97.0;' got
  the raw 32-bit float bit pattern instead of 99.
- size_t/wchar_t pinned to pointer-width / 16-bit so libc stack offsets
  stay where the asm functions were assembled for.  string.h's memset
  c arg is also 'unsigned char' instead of 'int' for the same reason.
- %lX (and %X under --int=32) honors width/zero-pad now: 0x68 with
  %04X prints "0068" instead of "68".
- run_ctests: per-variant expected-file lookup with prefix fallback
  (00178.int32.long64 → 00178.int32 → 00178); added int=32 expected
  for 00178; relaxed Z80-adapted 00212 to accept either int width.

c-testsuite --int=32: 211 → 217 (+6)
c-testsuite --int=32 --long=64: 211 → 217 (+6)
c-testsuite default: 219/220 (no regression)

+++ lib/include/wchar.h
@@ -10,10 +10,12 @@
 #include <stdarg.h>  /* for va_list */
 #include <stdio.h>   /* for FILE */
 
-/* Wide character type (if not already defined) */
+/* Wide character type (if not already defined).  Pinned to 16 bits so that
+ * L"..." literals (always emitted as 2 bytes per char by the codegen) match
+ * the array element size under both --int=16 and --int=32. */
 #ifndef _WCHAR_T_DEFINED
 #define _WCHAR_T_DEFINED
-typedef unsigned int wchar_t;
+typedef unsigned short wchar_t;
 #endif
 
 /* wint_t - type that can hold any wchar_t value plus WEOF */
